refactor: rename folder createClaimData to claimData
